fix(accounting): address ADA-360 PR review round 1
- Count MTD write-offs for all customers before balance filter
- Cast confirmed-order exposure sum to int in Drizzle adapter
- Add bulk SQL org read port; real Drizzle balances list query
- Extend as-of tests for postedAt and adjustment created_at
- Remove dead exposure code from business summary use case

+++ packages/sales/src/adapters/drizzle-open-order-exposure-read.ts
@@ -13,7 +13,7 @@ export class DrizzleOpenOrderExposureReadAdapter implements IOpenOrderExposureRe
   ): Promise<number> {
     const [row] = await this.db
       .select({
-        totalCents: sql<number>`coalesce(sum(${orderLines.qty} * ${orderLines.unitPriceCents}), 0)`,
+        totalCents: sql<number>`coalesce(sum((${orderLines.qty} * ${orderLines.unitPriceCents})::bigint), 0)::int`,
       })
       .from(orders)
       .innerJoin(orderLines, eq(orderLines.orderId, orders.id))
